Rename Aggregator to Buffer
