+++ src/aggregations/mod.rs
@@ -1,11 +1,9 @@
-mod aggregator;
 mod buffer;
 mod container;
 mod regex;
 mod systemdetails;
 mod whitelist;
 
-pub use self::aggregator::*;
 pub use self::buffer::*;
 pub use self::container::*;
 pub use self::regex::*;
